+++ frontend/src/Oauth/Authorize.jsx
@@ -12,7 +12,7 @@ export function Authorize() {
   const organizationId = authenticationService.getLoginOrganizationId();
 
   useEffect(() => {
-    authenticationService.clearUser();
+    !organizationId && authenticationService.clearUser();
     const errorMessage = router.query.error_description || router.query.error;
 
     if (errorMessage) {
