+++ frontend/src/_helpers/handle-response.js
@@ -17,3 +17,14 @@ export function handleResponse(response) {
     return data;
   });
 }
+export function handleResponseWithoutValidation(response) {
+  return response.text().then((text) => {
+    const data = text && JSON.parse(text);
+    if (!response.ok) {
+      const error = (data && data.message) || response.statusText;
+      return Promise.reject({ error, data });
+    }
+
+    return data;
+  });
+}
